requirements.txt
geographiclib==1.50
geopy==2.1.0
matplotlib==3.3.3
numpy==1.19.5
pandas==1.2.0
plotly==4.14.2
streamlit==0.74.1
pillow
requests
beautifulsoup4
